refactored trigger logic and added interval trigger logic
