+++ django_server/farminsight_dashboard_backend/action_scripts/__init__.py
@@ -1,4 +1,4 @@
 from .typed_action_script_factory import TypedActionScriptFactory
 from .typed_action_script import TypedSensor
-from .http_action_scripts import PostHttpActionScript
-from .tapo_p100_smart_plug_action_script import TapoP100SmartPlugActionScript
+from .tapo_p100_action_script import TapoP100SmartPlugActionScriptWithDelay
+from .shelly_plug_s_http_action_script import ShellyPlugHttpActionScript
