+++ django_server/farminsight_dashboard_backend/services/action_queue_services.py
@@ -1,11 +1,9 @@
 from django.utils.timezone import now
 
-from farminsight_dashboard_backend.models import ActionQueue, ControllableAction, ActionTrigger
-from farminsight_dashboard_backend.serializers import ActionQueueSerializer
-from farminsight_dashboard_backend.services.MeasurementTriggerManager import MeasurementTriggerManager
+from farminsight_dashboard_backend.models import ActionQueue
 
-from farminsight_dashboard_backend.services.action_trigger_services import get_action_trigger, get_all_auto_timeOfDay_action_triggers
-from farminsight_dashboard_backend.services.trigger_handlers import TriggerHandlerFactory
+from farminsight_dashboard_backend.services.action_trigger_services import get_all_active_auto_triggers
+from farminsight_dashboard_backend.services.trigger.trigger_handler_factory import TriggerHandlerFactory
 from farminsight_dashboard_backend.utils import get_logger
 from farminsight_dashboard_backend.action_scripts import TypedActionScriptFactory
 
@@ -59,11 +57,11 @@ def process_action_queue():
 
         # If the trigger action is already active skip it.
         # Cancel it. This should never happen anyway, actually.
-        if not is_new_action(action.id, trigger.id):
-            logger.info(f"Trigger '{trigger}' action is already active for this action.")
-            queue_entry.endedAt = now()
-            queue_entry.save()
-            continue
+        #if not is_new_action(action.id, trigger.id):
+        #    logger.info(f"Trigger '{trigger}' action is already active for this action.")
+        #    queue_entry.endedAt = now()
+        #    queue_entry.save()
+        #    continue
 
         # Execute the action
         try:
@@ -82,95 +80,19 @@ def process_action_queue():
         except Exception as e:
             logger.error(f"Failed to execute action {action.id}: {str(e)}")
 
-
 def create_auto_triggered_actions_in_queue(action_id=None):
-    """
-    This function will be called periodically and checks for all existing, active auto-triggers if they are triggering.
-    Trigger Type Logic will be called to check if it triggers and
-    the action queue will be checked that this trigger is not already active.
-    If a trigger triggers, an entry in the action queue will be created.
-    After all auto triggers are processed, process the action queue to start the actions.
-
-    An action_id can be passed to call the process only for one action. This will happen when the user actively selects the
-    "AUTO" button in the frontend.
-    :return:
-    """
     try:
-        auto_triggers = get_all_auto_timeOfDay_action_triggers(action_id)
+        auto_triggers = get_all_active_auto_triggers(action_id)
 
         for auto_trigger in auto_triggers:
-            # Trigger type logic to check for triggering && currently active trigger for the action must not be this trigger.
             handler = TriggerHandlerFactory.get_handler(auto_trigger)
-            if handler.should_trigger() and auto_trigger.action.isAutomated:
-                if is_new_action(auto_trigger.action.id, auto_trigger.id):
-                    serializer = ActionQueueSerializer(data={
-                                    "actionId": str(auto_trigger.action.id),
-                                    "actionTriggerId": str(auto_trigger.id),
-                                    "value": str(auto_trigger.actionValue),
-                                }, partial=True)
-                    if serializer.is_valid(raise_exception=True):
-                            serializer.save()
-                            logger.info(f"Queued auto trigger {auto_trigger.id} for action {auto_trigger.action.id}")
+            handler.enqueue_if_needed()
 
         process_action_queue()
 
     except Exception as e:
         logger.error(f"Failed to add action {action_id}: {str(e)}")
 
-
-def create_measurement_auto_triggered_actions_in_queue(sensor_id, measurement_value):
-    """
-    Creates an entry for the measurement auto trigger.
-    :param measurement_value:
-    :param sensor_id:
-    :return:
-    """
-
-    trigger_ids = MeasurementTriggerManager.get_trigger_ids_for_sensor(sensor_id)
-
-    if len(trigger_ids) >= 0:
-        for trigger_id in trigger_ids:
-            trigger = get_action_trigger(str(trigger_id))
-            # Trigger type logic to check for triggering && currently active trigger for the action must not be this trigger.
-            handler = TriggerHandlerFactory.get_handler(trigger)
-            if trigger.action.isAutomated and handler.should_trigger(measurement=measurement_value):
-                if is_new_action(trigger.action.id, trigger.id):
-                    serializer = ActionQueueSerializer(data={
-                        "actionId": str(trigger.action.id),
-                        "actionTriggerId": str(trigger.id),
-                        "value": str(measurement_value)
-                    }, partial=True)
-                    if serializer.is_valid(raise_exception=True):
-                        serializer.save()
-                        logger.info(f"Queued auto trigger {trigger.id} for action {trigger.action.id}")
-
-        process_action_queue()
-
-
-def create_manual_triggered_action_in_queue(action_id, trigger_id):
-    """
-    When the user manually selects a manual button in the frontend, the trigger will be activated and
-    an entry in the action queue will be created. (If no other actions for the same controllable actions are currently
-    running.
-    :param action_id:
-    :param trigger_id:
-    :return:
-    """
-    try:
-        trigger = get_action_trigger(trigger_id)
-        if trigger.isActive and is_new_action(action_id, trigger.id):
-            serializer = ActionQueueSerializer(data={
-                "actionId": action_id,
-                "actionTriggerId": trigger_id,
-            }, partial=True)
-            if serializer.is_valid(raise_exception=True):
-                serializer.save()
-                logger.debug(f"Queued auto trigger {trigger_id} for action {action_id}")
-
-        process_action_queue()
-
-    except Exception as e:
-        logger.error(f"Failed to add action {action_id}: {str(e)}")
 
 def get_active_state(controllable_action_id: str):
     """
@@ -185,7 +107,6 @@ def get_active_state(controllable_action_id: str):
     if latest_entry:
         return latest_entry.trigger
     return None
-
 
 def is_new_action(action_id, trigger_id):
     """
