+++ django_server/farminsight_dashboard_backend/services/action_trigger_services.py
@@ -14,29 +14,9 @@ def create_action_trigger(action_trigger_data:dict) -> ActionTriggerSerializer.d
 
     return serializer.save()
 
-
 def get_action_trigger(action_trigger_id):
     return get_object_or_404(ActionTrigger, pk=action_trigger_id)
 
-def get_all_auto_action_triggers(action_id):
-    """
-    Returns all active, not-manual triggers.
-    If an action_id is provided, return only the triggers of the action.
-    :return:
-    """
-    if not action_id:
-        return ActionTrigger.objects.filter(
-            isActive=True
-        ).exclude(
-            type='manual'
-        )
-    else:
-        return ActionTrigger.objects.filter(
-            isActive=True,
-            action_id=action_id
-        ).exclude(
-            type='manual'
-        )
 
 def get_all_auto_timeOfDay_action_triggers(action_id):
     """
@@ -54,4 +34,24 @@ def get_all_auto_timeOfDay_action_triggers(action_id):
             isActive=True,
             action_id=action_id,
             type='timeOfDay'
+        )
+
+def get_all_auto_interval_triggers(action_id=None):
+    if action_id:
+        return ActionTrigger.objects.filter(action__id=action_id, type="interval", isActive=True)
+    return ActionTrigger.objects.filter(type="interval", isActive=True)
+
+def get_all_active_auto_triggers(action_id=None):
+    if not action_id:
+        return ActionTrigger.objects.filter(
+            isActive=True
+        ).exclude(
+            type='manual'
+        )
+    else:
+        return ActionTrigger.objects.filter(
+            isActive=True,
+            action_id=action_id
+        ).exclude(
+            type='manual'
         )
