+++ django_server/farminsight_dashboard_backend/services/measurement_services.py
@@ -1,6 +1,7 @@
 from farminsight_dashboard_backend.models import Sensor
 
 
+
 def store_measurements_in_influx(sensor_id, data):
     from farminsight_dashboard_backend.services import InfluxDBManager
     sensor = Sensor.objects.get(id=sensor_id)
@@ -8,7 +9,8 @@ def store_measurements_in_influx(sensor_id, data):
                                                              sensor_id=sensor.id,
                                                              measurements=data)
     # Check for triggers paired to the sensor
-    from farminsight_dashboard_backend.services import create_measurement_auto_triggered_actions_in_queue
+    from farminsight_dashboard_backend.services.trigger.measurement_trigger_handler import \
+        create_measurement_auto_triggered_actions_in_queue
 
     last_measurement = data[len(data)-1]['value']
     create_measurement_auto_triggered_actions_in_queue(sensor_id, last_measurement)
