+++ django_server/farminsight_dashboard_backend/services/trigger/MeasurementTriggerManager.py
